chore: Refactor package names
